+++ src/styles/components/TaskCenter.css
@@ -1,0 +1,53 @@
+/* 搜索框样式 */
+.task-search .ant-input-search-button {
+  background: #1890ff;
+  border-color: #1890ff;
+  height: 40px;
+  width: 40px;
+  border-radius: 0 8px 8px 0;
+}
+
+.task-search .ant-input-search-button:hover {
+  background: #40a9ff;
+  border-color: #40a9ff;
+}
+
+.task-search .ant-input-affix-wrapper {
+  border-radius: 8px 0 0 8px;
+  height: 40px;
+  border-right: none;
+  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.06);
+}
+
+.task-search .ant-input-affix-wrapper:hover,
+.task-search .ant-input-affix-wrapper:focus {
+  border-color: #40a9ff;
+}
+
+.task-search .ant-input-affix-wrapper-focused {
+  box-shadow: 0 0 0 2px rgba(24, 144, 255, 0.2);
+}
+
+/* 发布任务按钮样式 */
+.publish-task-btn {
+  height: 40px !important;
+  padding: 0 24px !important;
+  border-radius: 8px !important;
+  background: #1890ff !important;
+  border: none !important;
+  display: flex !important;
+  align-items: center !important;
+  gap: 8px !important;
+  transition: all 0.3s ease !important;
+  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.06) !important;
+}
+
+.publish-task-btn:hover {
+  background: #40a9ff !important;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1) !important;
+  transform: translateY(-1px);
+}
+
+.publish-task-btn .anticon {
+  font-size: 16px;
+}
